# Message.
It didn't work the first time, but I believe I fixed it after familiarising myself with the SQL Command syntax. 'SET' specifies what the name needs to be changed to; although I was on the right path, In the Storage Manager, I hadn't created a seperate string to declare as the new name under the artistName field, part of the command. It was working at first, but I was renaming the field with its current name; redundant.

I found it hard to understand the syntax at first, but I realised my Delete command deleted by name; which made me realise I just had to tweak the SQL to match it.