I might be onto something. For the 'go back' system; I moved the whole batch of switch cases related to the menuscreen into a Method seperate from the Main Method; Fixed some of it, but other cases do not work at the moment.
The purpose?
Something to do with the bool value used by the do while in the go back system. When I would call the Administrator Menu inside of the case of 'y', it wouldn't loop back around because the calling of the Administrator Menu became a seperate, new instance of itself, and was still happening inside the case of the 'go back' where 'y' was a valid input, thus it didn't loop back into the external do while of the options to choose from. Creating a seperate Method for the menuscreen options avoids this. I will offer a better explanation in my iterative improvement.